+++ ssd/ssd.vcxproj.filters
@@ -30,5 +30,8 @@
     <ClCompile Include="SSDManager.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
+    <ClCompile Include="Exception.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
   </ItemGroup>
 </Project>
